- Marca de agua en PDFs - Precios en decimal - Descripción de artículos - Modificar nombre, precio y unidad desde app - Arreglado bug fecha UTC

+++ FacturacionDobleEje/Models/Quote.cs
@@ -11,11 +11,11 @@ namespace FacturacionDobleEje.Models
         public DateTime Date { get; set; } = DateTime.Now;
         public required Client Client { get; set; }
         public List<QuoteLine> Lines { get; set; } = new List<QuoteLine>();
-        public double VatType { get; set; } = 0.21;
+        public decimal VatType { get; set; } = 0.21m;
 
-        public double Subtotal => Lines.Sum(l => l.Amount);
-        public double Vat => Math.Round(Subtotal * VatType, 2);
-        public double Total => Subtotal + Vat;
+        public decimal Subtotal => Lines.Sum(l => l.Amount);
+        public decimal Vat => Math.Round(Subtotal * VatType, 2);
+        public decimal Total => Subtotal + Vat;
         public required string Status { get; set; }
         public DateTime CreatedOn { get; set; } = DateTime.UtcNow;
     }
